# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "@aasumitro",
            "url": "https://aasumitro.id/",
            "email": "hello@aasumitro.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/aasumitro/posbe/blob/main/LICENSE"
        }
    },
    "paths": {
...
        "/v1/rooms": {
            "get": {
                "description": "Get Room List.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rooms"
                ],
                "summary": "Room List",
                "responses": {
                    "200": {
                        "description": "OK RESPOND",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessRespond"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.Room"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    }
                }
            },
            "post": {
                "description": "Create new Room.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rooms"
                ],
                "summary": "Store Room Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "floor id",
                        "name": "floor_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "x position",
                        "name": "x_pos",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "y position",
                        "name": "y_pos",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "weight",
                        "name": "w_size",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "height",
                        "name": "h_size",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "capacity",
                        "name": "capacity",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "price",
                        "name": "price",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "CREATED RESPOND",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessRespond"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Room"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE ENTITY RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ValidationErrorRespond"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    }
                }
            }
        },
        "/v1/rooms/{id}": {
            "put": {
                "description": "Update Room Data by ID.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rooms"
                ],
                "summary": "Update Room Data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "room id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "floor id",
                        "name": "floor_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "x position",
                        "name": "x_pos",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "y position",
                        "name": "y_pos",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "weight",
                        "name": "w_size",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "height",
                        "name": "h_size",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "capacity",
                        "name": "capacity",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "price",
                        "name": "price",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CREATED RESPOND",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessRespond"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Room"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "BAD REQUEST RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE ENTITY RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ValidationErrorRespond"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Room Data by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rooms"
                ],
                "summary": "Delete Room Data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "room id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "NO CONTENT RESPOND"
                    },
                    "400": {
                        "description": "BAD REQUEST RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Room": {
            "type": "object",
            "required": [
                "capacity",
                "floor_id",
                "h_size",
                "name",
                "price",
                "w_size",
                "x_pos",
                "y_pos"
            ],
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "created_at": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "floor_id": {
                    "type": "integer"
                },
                "h_size": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "updated_at": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "w_size": {
                    "type": "number"
                },
                "x_pos": {
                    "type": "number"
                },
                "y_pos": {
                    "type": "number"
                }
            }
        },
...
        "sql.NullInt64": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer"
                },
                "valid": {
                    "description": "Valid is true if Int64 is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "utils.ErrorRespond": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "utils.SuccessRespond": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        },
        "utils.ValidationErrorRespond": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        }
    }
}